==> simulations/09_other_variants.ipynb <==
# Try a new method

#import shap
import pandas as pd
import scipy as sp
import numpy as np
#from irf.ensemble import wrf
from sklearn.ensemble import RandomForestClassifier as rfc
from sklearn.ensemble import RandomForestRegressor as rfr
from sklearn.preprocessing import LabelEncoder, OneHotEncoder, LabelBinarizer, MinMaxScaler
from sklearn.metrics import accuracy_score, roc_auc_score, f1_score
import matplotlib.pyplot as plt
from treeinterpreter.feature_importance import feature_importance as FI
from treeinterpreter.feature_importance import MDA
import sys
sys.path += ['../code/']

### method 1: demean MDI-oob

### load data

num = 2

problem = [0]
tree_depth = [1]
for num in range(1, 3):
    for problem in ['classification', 'regression']:
        for tree_depth in ['deep', 'shallow']:
            #if tree_depth == 'deep':
            #    continue
            #if (num == 1 and tree_depth == 'deep' and problem == 'regression'):
            #    continue
            print('working on {} {} {}'.format(num, problem, tree_depth))
            gini_list, demean_list = [], []
            for i in range(40):
                print(i)
                X_train = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_X_train.csv'.format(problem, num, i), delimiter=',')
                y_train = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_y_train.csv'.format(problem, num, i), delimiter=',')
                X_test = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_X_test.csv'.format(problem, num, i), delimiter=',')
                y_test = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_y_test.csv'.format(problem, num, i), delimiter=',')
                noisy_features = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_noisy_features.csv'.format(problem, num, i), delimiter=',').astype(np.int)
                n_features = len(noisy_features)

                ### Train RF
                if problem == 'classification':
                    if tree_depth == 'deep':
                        rf = rfc(n_estimators=100, max_features=10, min_samples_leaf=1)
                    else:
                        rf = rfc(n_estimators=100, max_features=10, min_samples_leaf=100)
                else:
                    if tree_depth == 'deep':
                        rf = rfr(n_estimators=100, max_features= 10, min_samples_leaf=1)
                    else:
                        rf = rfr(n_estimators=100, max_features= 10, min_samples_leaf=100)
                rf.fit(X_train, y_train)
                gini_imp = rf.feature_importances_
                if problem == 'classification':
                    y_train_one_hot = OneHotEncoder().fit_transform(y_train[:, np.newaxis]).todense()
                    demean_imp, debiased_fi_std = FI(rf, X_train, y_train_one_hot, type='oob', normalized=False, balanced = False, demean=True)
                else:
                    demean_imp, debiased_fi_std = FI(rf, X_train, y_train[:, np.newaxis], type='oob', normalized=False, balanced = False, demean=True)
                #ufi_imp = UFI_importance(rf, X_train, y_train)
                #ufi_imp[ufi_imp < 0] = 0
                gini_list.append(roc_auc_score(noisy_features, - gini_imp))
                #demean_imp = FI(rf, X_train, y_train[:,np.newaxis], demean=True)
                #ufi_list.append(roc_auc_score(noisy_features, - ufi_imp))
                demean_list.append(roc_auc_score(noisy_features, -demean_imp))
                
            np.savez('../intermediate/04_aggregate/{}{}/add_new_demean_results_{}.npz'.format(problem, num, tree_depth), gini_list=gini_list, demean_list=demean_list)


### method 2 : standardize contribution

### load data

num = 2

problem = [0]
tree_depth = [1]
for num in range(1, 3):
    for problem in ['classification', 'regression']:
        for tree_depth in ['deep', 'shallow']:
            #if tree_depth == 'deep':
            #    continue
            #if (num == 1 and tree_depth == 'deep' and problem == 'regression'):
            #    continue
            print('working on {} {} {}'.format(num, problem, tree_depth))
            gini_list, demean_list = [], []
            for i in range(40):
                print(i)
                X_train = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_X_train.csv'.format(problem, num, i), delimiter=',')
                y_train = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_y_train.csv'.format(problem, num, i), delimiter=',')
                X_test = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_X_test.csv'.format(problem, num, i), delimiter=',')
                y_test = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_y_test.csv'.format(problem, num, i), delimiter=',')
                noisy_features = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_noisy_features.csv'.format(problem, num, i), delimiter=',').astype(np.int)
                n_features = len(noisy_features)

                ### Train RF
                if problem == 'classification':
                    if tree_depth == 'deep':
                        rf = rfc(n_estimators=100, max_features=10, min_samples_leaf=1)
                    else:
                        rf = rfc(n_estimators=100, max_features=10, min_samples_leaf=100)
                else:
                    if tree_depth == 'deep':
                        rf = rfr(n_estimators=100, max_features= 10, min_samples_leaf=1)
                    else:
                        rf = rfr(n_estimators=100, max_features= 10, min_samples_leaf=100)
                rf.fit(X_train, y_train)
                gini_imp = rf.feature_importances_
                if problem == 'classification':
                    y_train_one_hot = OneHotEncoder().fit_transform(y_train[:, np.newaxis]).todense()
                    demean_imp, debiased_fi_std = FI(rf, X_train, y_train_one_hot, type='oob', normalized=False, balanced = False, normal_fX = True)
                else:
                    demean_imp, debiased_fi_std = FI(rf, X_train, y_train[:, np.newaxis], type='oob', normalized=False, balanced = False, normal_fX=True)
                #ufi_imp = UFI_importance(rf, X_train, y_train)
                #ufi_imp[ufi_imp < 0] = 0
                gini_list.append(roc_auc_score(noisy_features, - gini_imp))
                #demean_imp = FI(rf, X_train, y_train[:,np.newaxis], demean=True)
                #ufi_list.append(roc_auc_score(noisy_features, - ufi_imp))
                demean_list.append(roc_auc_score(noisy_features, -demean_imp))
                
            np.savez('../intermediate/04_aggregate/{}{}/add_new_norm_results_{}.npz'.format(problem, num, tree_depth), gini_list=gini_list, norm_list=demean_list)



### look at some results

problem = 'classification'
tree_depth = 'shallow'
num = 2
i = 10
X_train = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_X_train.csv'.format(problem, num, i), delimiter=',')
y_train = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_y_train.csv'.format(problem, num, i), delimiter=',')
X_test = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_X_test.csv'.format(problem, num, i), delimiter=',')
y_test = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_y_test.csv'.format(problem, num, i), delimiter=',')
noisy_features = np.loadtxt('../intermediate/04_aggregate/{}{}/permuted{}_noisy_features.csv'.format(problem, num, i), delimiter=',').astype(np.int)

rf = rfc(n_estimators=100, max_features=10, min_samples_leaf=1)
rf.fit(X_train, y_train)

%timeit naive_imp = naive_OOB_importance(rf, X_train, y_train)

np.mean(naive_list), np.mean(gini_list)

roc_auc_score(noisy_features, -naive_imp), roc_auc_score(noisy_features, -rf.feature_importances_)
